Refactor global-teardown.ts
